rename file to cesium_properties_widget_builder
