+++ exts/cesium.omniverse/cesium/omniverse/ui/attributes/web_map_service_raster_overlay_attributes_widget.py
@@ -4,7 +4,7 @@ from omni.kit.property.usd.usd_property_widget import SchemaPropertiesWidget
 from cesium.usd.plugins.CesiumUsdSchemas import (
     WebMapServiceRasterOverlay as CesiumWebMapServiceRasterOverlay,
 )
-from .attribute_widget_util import build_slider, build_common_raster_overlay_properties
+from .cesium_properties_widget_builder import build_slider, build_common_raster_overlay_properties
 
 
 class CesiumWebMapServiceRasterOverlayAttributesWidget(SchemaPropertiesWidget):
